+++ app/models/supplier.rb
@@ -1,10 +1,13 @@
 class Supplier < ApplicationRecord
   has_many :agreements, dependent: :destroy
   has_many :frameworks, through: :agreements
+  has_many :active_frameworks, -> { merge(Agreement.active) }, through: :agreements, class_name: 'Framework',
+           source: :framework
   has_many :submissions, dependent: :nullify
   has_many :tasks, inverse_of: :supplier, dependent: :destroy
   has_many :memberships, dependent: :destroy
   has_many :users, through: :memberships
+  has_many :active_users, -> { active }, through: :memberships, class_name: 'User', source: :user
 
   validates :name, presence: true
   validates :coda_reference, allow_nil: true, format: {