+++ config/initializers/date_format.rb
@@ -1,2 +1,3 @@
 Date::DATE_FORMATS[:month_year] = '%B %Y'
+Date::DATE_FORMATS[:day_month_year] = '%-d %B %Y'
 Time::DATE_FORMATS[:date_with_utc_time] = '%e %B %Y %H:%M %Z'